rlTime V2 made capable of holding relative time values.
Assertions and type conversion rules ensure to combine only valid
types of time (only one absolute time in addition or subtraction, all
others need to be relative time values)
